+++ src/stories/element/TextArea.jsx
@@ -9,3 +9,7 @@ const WrapperTextArea = elementWrapper(TextArea);
 storiesOf('Element.TextArea', module).add('Empty', () => (
   <WrapperTextArea value="" name="test-input" onChange={action('onChange')} />
 ));
+
+storiesOf('Element.TextArea', module).add('With content', () => (
+  <WrapperTextArea value="<p>Hello there</p>" name="test-input" onChange={action('onChange')} />
+));